updated the tests for base table dataset and managed table dataset
